requirements.txt
async-timeout==4.0.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect==1.2.0
cssutils==2.9.0
exceptiongroup==1.2.0
fastapi==0.105.0
iniconfig==2.0.0
loguru==0.7.2
lxml==4.9.3
packaging==23.2
pika==1.3.2
pluggy==1.3.0
premailer==3.10.0
pydantic==2.5.2
pydantic_core==2.14.5
pytest==7.4.3
pytest-mock==3.12.0
python-dotenv==1.0.0
requests==2.31.0
slack-sdk==3.26.1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.24.0.post1
pyyaml==6.0.1
httpx==0.25.2
httpcore==1.0.2